Move visualization to dedicated module
