+++ acrosome_counter/predictor.py
@@ -8,33 +8,18 @@ import matplotlib as mpl
 from matplotlib import pyplot as plt
 from detectron2.engine import DefaultPredictor
 from detectron2.data import MetadataCatalog
-from detectron2.utils.visualizer import Visualizer
 import pandas as pd
 
-from acrosome_counter.inputs import MAP_IDS, MAP_NAMES
+from acrosome_counter.inputs import MAP_NAMES
+from acrosome_counter.visualize import visualize
 
 mpl.use('TkAgg')
-
-CLASSES_COLORS = [(0, 1, 0), (1, 1, 1), (0, 0, 1)]
-
-
-class Visualizer(Visualizer):
-    def overlay_instances(self, **kwargs):
-        labels = [label.split(" ")[0] for label in kwargs['labels']]
-        percentages = [label.split(" ")[1] for label in kwargs['labels']]
-        category_ids = [MAP_IDS[label] for label in labels]
-        kwargs['assigned_colors'] = [
-            self.metadata.thing_colors[c] for c in category_ids
-        ]
-        kwargs['labels'] = percentages
-        return super().overlay_instances(**kwargs)
 
 
 class Predictor(DefaultPredictor):
     def __init__(self, cfg):
         super().__init__(cfg)
         self.metadata = MetadataCatalog.get("test")
-        self.metadata.thing_colors = CLASSES_COLORS
         self.results = {}
         self.dataset = None
 
@@ -47,18 +32,8 @@ class Predictor(DefaultPredictor):
             input_image = image[..., [2, 1]]
             outputs = super().__call__(input_image)
             self.results[image_path] = outputs['instances']
-            visualizer = Visualizer(
-                image,
-                metadata=self.metadata,
-                scale=3.0,
-            )
-            out = visualizer.draw_instance_predictions(
-                outputs["instances"].to("cpu")
-            )
-            annotated_image = out.get_image()
             if plot:
-                plt.imshow(annotated_image)
-                plt.show()
+                visualize(image, outputs, self.metadata)
 
     def export_xml(self, dest_path=None):
         if dest_path is None:
